% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.nloptr.R
\name{is.nloptr}
\alias{is.nloptr}
\title{R interface to NLopt}
\usage{
is.nloptr(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
Logical. Return TRUE if all tests were passed, otherwise return
FALSE or exit with Error.
}
\description{
is.nloptr preforms checks to see if a fully specified problem is supplied to
nloptr. Mostly for internal use.
}
\seealso{
\code{\link[nloptr:nloptr]{nloptr}}
}
\author{
Jelmer Ypma
}
\keyword{interface}
\keyword{optimize}
