% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{nl.grad}
\alias{nl.grad}
\alias{nl.jacobian}
\title{Numerical Gradients and Jacobians}
\usage{
nl.grad(x0, fn, heps = NULL, points = 3, ...)
}
\arguments{
\item{x0}{Point as a vector where the gradient is to be calculated.}

\item{fn}{Scalar function of one or several variables.}

\item{heps}{Step size to be used. Defaults to \code{NULL} and is selected
based on the number of points used in the estimation algorithm.}

\item{points}{Either a 3 or 5 point numerical estimate.}

\item{\dots}{Additional arguments passed to the function.}
}
\value{
\code{grad} returns the gradient as a vector;
\code{jacobian} returns the Jacobian as a matrix of the usual dimensions.
}
\description{
Provides numerical gradients and Jacobians.
}
\details{
Both functions apply a \dQuote{central difference formula} with the step size
selected as recommended in the literature depending on whether a three or
five point estimate is used.
}
\examples{

  fn1 <- function(x) sum(x ^ 2)
  nl.grad(seq(0, 1, by = 0.2), fn1)
  ## [1] 0.0  0.4  0.8  1.2  1.6  2.0
  nl.grad(rep(1, 5), fn1)
  ## [1] 2  2  2  2  2

  fn2 <- function(x) c(sin(x), cos(x))
  x <- (0:1) * 2 * pi
  nl.jacobian(x, fn2, points = 5)
  ##    [,1] [,2]
  ## [1,]  1  0
  ## [2,]  0  1
  ## [3,]  0  0
  ## [4,]  0  0

}
\references{
Sauer, Timothy (2012) \emph{Numerical Analysis} (2nd ed.).
Addison-Wesley Publishing Company. ISBN 978-0321783677
}
\author{
Hans W. Borchers, Avraham Adler \email{Avraham.Adler@gmail.com}
}
